requirements.txt
soccerdata>=1.8.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
python-telegram-bot>=20.0
schedule>=1.2.0
python-dotenv>=1.0.0
joblib>=1.3.0
pyarrow>=12.0.0
requests>=2.31.0
pytest>=7.0.0
